+++ example/src/PipesExamples.component.ts
@@ -1,25 +1,25 @@
-import { Component } from '@angular/core';
-import { fxArray } from 'ack-angular-fx'
+import { Component } from "@angular/core";
+import { fxArray } from "ack-angular-fx"
 
 import {
   PageScrollService,
   PageScrollInstance
-} from 'ng2-page-scroll'
+} from "ngx-page-scroll"
 
-import {string as pipesExamples} from './templates/pipes-examples.pug'
+import {string as pipesExamples} from "./templates/pipes-examples.pug"
 @Component({
-  selector: 'pipes-examples'
+  selector: "pipes-examples"
   ,template: pipesExamples
   ,animations:fxArray
 }) export class PipesExamples {
   views:any = {}
-  simpleArray = ['a','b','c']
+  simpleArray = ["a","b","c"]
 
   constructor(public PageScrollService:PageScrollService){}
 
   scrollToModuleImport(){
     setTimeout(()=>{
-      const pageScrollInstance = PageScrollInstance.simpleInstance(document, '#Import AckModule');
+      const pageScrollInstance = PageScrollInstance.simpleInstance(document, "#Import AckModule");
       this.PageScrollService.start(pageScrollInstance);
     }, 600)
   }