+++ src/AckModule.ts
@@ -1,11 +1,11 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule }   from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { NgModule } from "@angular/core"
+import { HttpClientModule } from "@angular/common/http"
+import { CommonModule } from "@angular/common"
+import { FormsModule }   from "@angular/forms"
 
-import { providers } from "./providers";
-import { declarations as components } from "./declarations";
-import { declarations as pipes } from "./pipes";
+import { providers } from "./providers"
+import { declarations as components } from "./declarations"
+import { declarations as pipes } from "./pipes"
 
 const declarations = [...components, ...pipes]
 
@@ -14,18 +14,17 @@ const declarations = [...components, ...pipes]
   imports:[
     CommonModule
     ,FormsModule
-    ,HttpModule
-    //,HttpClientModule
+    //,HttpModule
+    ,HttpClientModule
     //,AckOffline
     //,BrowserAnimationsModule
-    //,Ng2PageScrollModule.forRoot()
   ]
   ,declarations: declarations
   ,providers:providers,
   exports:[
     FormsModule,
-    HttpModule,
-    //HttpClientModule,
+    //HttpModule,
+    HttpClientModule,
     ...declarations,
    ]
 }) export class AckModule {}