+++ src/providers/ErrorLog.ts
@@ -30,11 +30,11 @@ import { Injectable } from '@angular/core';
     return logObToErrorObject(ob)
   }
 
-  //api error to js Error object
-  logObToErrorObject(ob){
+  /*logObToErrorObject(ob){
     return logObToErrorObject(ob)
-  }
+  }*/
 
+  //api error to js Error object
   paramAudit(e, toConsole?){
     if(e.constructor==String){
       e = new Error(e)