+++ test/app.module.ts
@@ -1,9 +1,9 @@
-import { NgModule, Component } from '@angular/core'
-import { BrowserModule } from '@angular/platform-browser'
+import { NgModule, Component } from "@angular/core"
+import { BrowserModule } from "@angular/platform-browser"
 
 @Component({
-	selector: 'app',
-	template: 'nothing here'
+	selector: "app",
+	template: "nothing here"
 }) export class AppComponent {}
 
 @NgModule({